requirements.txt
pandas
pypistats
google-cloud-bigquery
matplotlib